requirements.txt
django == 3.2.7
djangorestframework==3.12.4
celery==5.0.5
flower==1.0.0
